backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
psycopg2-binary
asyncpg
python-dotenv
